Allow time inconsistencies in general when verifying an ID Token
I don't like this myself, but as we already have the "feature" implemented to allow Tokens that have been issued in the future, it's sadly only consequent to also allow when the "auth_time" claim is off in a similar fashion.

The next major release will have a better solution.